modified plugin file structure
